feat: Adds `isFocused` prop to Form inputs to statefully manage focus
Preact treats `autoFocus` a lot less aggresivelly, than react (more in
line with native dom behaviour) so this is a workaround for statefully
changing which element is focused.

This is behaviour is to support some UX for the add invitee list where
we want to auto focus the firstname field of the newly added invitee.